backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.5
Django==5.2.12
django-cors-headers==4.9.0
django-ninja==1.5.3
django-ninja-extra==0.31.0
django-ninja-jwt==5.4.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.0
injector==0.24.0
numpy==2.2.6
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.2
redis==7.3.0
rsa==4.9.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
Pillow==10.3.0
requests==2.32.3
channels==4.2.0
channels-redis==4.2.0
daphne==4.1.2